round up the number of batches per epoch
Instead of np.floor, I think np.ceil is the right form of rounding to use. this is with reference to the tensorflow/keras documentation (https://www.tensorflow.org/api_docs/python/tf/keras/utils/Sequence)

Example explanation:
given a sample size (or the length of list_enzymes in this context) of 35 and a batch_size of 10, using np.floor will result in 3 batches. This leaves out the remaining 5 samples in the list_enzymes. meanwhile, using the ceil will result in 4 batches which covers the last 5 samples. You may be wondering about indexing in __getitem__ but since the self.indexes is a numpy array, indexing beyond the length of that array will just truncate at the end of the arra.